+++ Assets/ELEMENTS/Runtime/Elements/WindowStartPosition.cs
@@ -1,0 +1,11 @@
+namespace ELEMENTS.Elements
+{
+    public enum WindowStartPosition
+    {
+        Center,
+        TopLeft,
+        TopRight,
+        BottomLeft,
+        BottomRight,
+    }
+}